function pointer

+++ operators.c
@@ -1,13 +1,13 @@
 #include "operators.h"
 
-int add(int op1, int op2) {
-	return op1+op2;
+double add(int op1, int op2) {
+	return (double)op1+op2;
 }
-int minus(int op1, int op2) {
-	return op1-op2;
+double minus(int op1, int op2) {
+	return (double)op1-op2;
 }
-int mul(int op1, int op2) {
-	return op1*op2;
+double mul(int op1, int op2) {
+	return (double)op1*op2;
 }
 
 double div(int op1, int op2) {
